Crate hopkarp

Finds a maximum cardinality bipartite matching by Hopcroft―Karp’s algorithm.

Basic usage

Use a function hopkarp and get results via HopkarpResult.

Example

use hopkarp::hopkarp;

let g = vec![vec![0, 1], vec![0]];
let result = hopkarp(2, &g);
assert_eq!(result.count, 2);

Structs

HopkarpResult

Summary of the result of Hopcroft―Karp’s algorithm.

Functions

hopkarp

Takes a forward adjacency list and the length of the right side, and returns a maximum cardinality bipartite matching.